+++ cmd/main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/infracloudio/firecracker-marathon/runtime"
+import "github.com/infracloudio/firecracker-marathon/pkg/runtime"
 
 func main() {
 	runtime.Start()
