+++ pkg/gateway/server.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/gorilla/mux"
-	"github.com/infracloudio/firecracker-marathon/logging"
+	"github.com/infracloudio/firecracker-marathon/pkg/logging"
 )
 
 func Start() {
