+++ pkg/runtime/server.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/gorilla/mux"
-	"github.com/infracloudio/firecracker-marathon/logging"
+	"github.com/infracloudio/firecracker-marathon/pkg/logging"
 	"github.com/songgao/water"
 )
